AND-3784: add external method to prepare hash to send
