+++ blockchain/src/main/java/com/tangem/blockchain/blockchains/ethereum/EthereumUtils.kt
@@ -1,11 +1,9 @@
 package com.tangem.blockchain.blockchains.ethereum
 
-import com.tangem.blockchain.common.Amount
-import com.tangem.blockchain.common.AmountType
-import com.tangem.blockchain.common.Blockchain
-import com.tangem.blockchain.common.TransactionData
+import com.tangem.blockchain.common.*
 import com.tangem.common.extensions.hexToBytes
 import com.tangem.common.extensions.toByteArray
+import com.tangem.common.extensions.toDecompressedPublicKey
 import org.kethereum.crypto.api.ec.ECDSASignature
 import org.kethereum.crypto.determineRecId
 import org.kethereum.crypto.impl.ec.canonicalise
@@ -109,6 +107,39 @@ class EthereumUtils {
                 .encode(SignatureData(v = chainId.toBigInteger()))
                 .keccak()
             return CompiledEthereumTransaction(transaction, hash)
+        }
+
+        fun prepareTransactionToSend(
+            signature: ByteArray,
+            transactionToSign: CompiledEthereumTransaction,
+            walletPublicKey: Wallet.PublicKey,
+            blockchain: Blockchain,
+        ): ByteArray {
+            val publicKey = walletPublicKey.blockchainKey.toDecompressedPublicKey().sliceArray(1..64)
+            return prepareTransactionToSend(signature, transactionToSign, publicKey, blockchain)
+        }
+
+        fun prepareTransactionToSend(
+            signature: ByteArray,
+            transactionToSign: CompiledEthereumTransaction,
+            walletPublicKey: ByteArray,
+            blockchain: Blockchain,
+        ): ByteArray {
+            val r = BigInteger(1, signature.copyOfRange(0, 32))
+            val s = BigInteger(1, signature.copyOfRange(32, 64))
+
+            val ecdsaSignature = ECDSASignature(r, s).canonicalise()
+
+            val recId = ecdsaSignature.determineRecId(
+                transactionToSign.hash,
+                PublicKey(walletPublicKey)
+            )
+            val chainId = blockchain.getChainId()
+                ?: error("${blockchain.fullName} blockchain is not supported by ${this::class.simpleName}")
+            val v = (recId + 27 + 8 + (chainId * 2)).toBigInteger() // EIP-155
+            val signatureData = SignatureData(ecdsaSignature.r, ecdsaSignature.s, v)
+
+            return transactionToSign.transaction.encode(signatureData)
         }
 
         fun buildApproveToSign(
